AppLoader now is iOS only, on android it renders the children
