+++ lib/components/AppLoader/AppLoaderIOS/index.d.ts
@@ -1,0 +1,2 @@
+import AppLoaderIOS from './AppLoaderIOS';
+export default AppLoaderIOS;
