+++ lib/components/AppLoader/AppLoaderIOS/index.js
@@ -1,0 +1,5 @@
+"use strict";
+Object.defineProperty(exports, "__esModule", { value: true });
+var AppLoaderIOS_1 = require("./AppLoaderIOS");
+exports.default = AppLoaderIOS_1.default;
+//# sourceMappingURL=index.js.map
