+++ lib/components/AppLoader/AppLoaderIOS/styles.js
