+++ lib/utils/platform.d.ts
@@ -1,0 +1,3 @@
+export declare function getPlatform(): import("react-native").PlatformOSType;
+export declare function isIOS(): boolean;
+export declare function isAndroid(): boolean;
