+++ src/components/AppLoader/AppLoaderIOS/index.ts
@@ -1,0 +1,2 @@
+import AppLoaderIOS from './AppLoaderIOS';
+export default AppLoaderIOS;
