+++ src/components/AppLoader/AppLoaderIOS/styles.ts
